requirements.txt
pillow
tqdm
termcolor